Spinoboxifying: enforce setting boundaries
Unfortunately, QT spinboxes have an arbitrary default maximum of ~100,
which can easily limit the true parameter range. To address this:
- demand explicit maximum (and minimum) in `create_spinbox` helper
- re-adjust maxima when restoring parameters #36

It's maybe still possible that there is a way to end up with parameter
values outside the spinbox boundaries, especially when users set
ridculous units like dt=10_000min But catching all of this now is
overkill. #35